New Calibration function
The calibration panel can be used to read in two calibration files for
both NOPAs and set the stages to compensate the GVD for a specific
wavelength (interpolating between the data in the calibration files).
If both NOPAs are used at the same time, the roofmirrors are set to
equalize both beam path lengths to achieve a temporal pulse overlap at
the sample.