Added the additional beam path for second harmonic mode
If one of the NOPAs is operated in the second harmonic mode, the beam
path within the NOPA itself is increased by around 60 cm.
If BOTH NOPAs are used in the second harmonic mode or fundamental mode,
the additional paths cancel each other out.
However, if only one NOPA is operated in the second harmonic mode, the
beam path of the other NOPA has to be increased as well.
This is achieved by adding additional 2*30 cm of beam path using the
roofmirror.
If a NOPA is used in second harmonic mode can be determined by the
selected wavelength in the calibration panel.